feat: add router flags

+++ packages/create-vue-vine/src/utils/renderTemplate.ts
@@ -1,5 +1,8 @@
 import { copyFile, mkdir, readdir, stat } from 'node:fs/promises'
 import path from 'node:path'
+import fs from 'node:fs'
+import sortDependencies from './sortDependencies'
+import deepMerge from './deepMerge'
 
 export async function renderTemplate(src: string, dest: string) {
   const stats = await stat(src)
@@ -17,6 +20,15 @@ export async function renderTemplate(src: string, dest: string) {
   if (filename.startsWith('_')) {
     // rename `_file` to `.file`
     dest = path.resolve(path.dirname(dest), `.${filename.slice(1)}`)
+  }
+
+  if (filename === 'package.json' && fs.existsSync(dest)) {
+    // merge instead of overwriting
+    const existing = JSON.parse(fs.readFileSync(dest, 'utf8'))
+    const newPackage = JSON.parse(fs.readFileSync(src, 'utf8'))
+    const pkg = sortDependencies(deepMerge(existing, newPackage))
+    fs.writeFileSync(dest, `${JSON.stringify(pkg, null, 2)}\n`)
+    return
   }
 
   await copyFile(src, dest)
